requirements.txt
h5py==3.11.0
huggingface-hub==0.17.2
hydra-core==1.3
numpy==1.26.0
opencv-python==4.10.0.84
pillow==10.3.0
pygame==2.5.2
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.4
wandb==0.17.0
matplotlib>=3.5.0
